+++ src/gpu/gpu_eri_grad_assembler_spdf_7_3.h
@@ -19,7 +19,7 @@
 #define VY(a,b,c) (YVerticalTemp[(c)])
 
 
-__device__ static inline void ERint_grad_vertical_spdf_7_3(uint8_t I, uint8_t J, uint8_t K, uint8_t L,
+__device__ static inline void ERint_grad_vertical_spdf_7_3(uint32_t I, uint32_t J, uint32_t K, uint32_t L,
         const QUICKDouble Ptempx, const QUICKDouble Ptempy, const QUICKDouble Ptempz,
         const QUICKDouble WPtempx, const QUICKDouble WPtempy, const QUICKDouble WPtempz,
         const QUICKDouble Qtempx, const QUICKDouble Qtempy, const QUICKDouble Qtempz,
